«Added new functions and worked with colors and header»
Added possibility to upload SVG, but it did not work when I tried to use a SVG logo file. The customized header works as inteded. The front page is not the front page, so it was right for a very long time. My logo became blurry when I switched it out with a new alteration. Have not fixed this yet. Added color variables to the headings.