«Header and menus»
Solved the issue with the blurred logo (but not sure how...), and implemented code for the main navigation from Morten Rand-Henriksens Lynda.com course based on the theme _s. Worked on the centered version (is_front_page()) and the others, but when I implemented the new menu, this must be re-done.