«Post navigation and single post»
I have the structure for my layout, but the flex will not kick in… very close to my goal for the blog posts template that will be used on the news page and the tourroutes page. It seems like my Sass neither will kick in, and that's weird, since it works like a charm in the browser inspector, but not when I do exactly the same in Dreamweaver… Frustrating. Other than that, removed sidebar / widget areas since my website not will display them.