«Home-page and galleries»
Used the plug-in Nextgen gallery and used the album function to create the different galleries on the gallery-page. I added custom CSS for this, but it needs more styling. The home-page, or the landing-page is finished, and I used the posts-and-pages.scss to style it (mostly). I added a function in the php file to remove the p-tag that is added automaticly to images that is placed on pages. Added a custom class to thumbnail images on the home-page, to make them rounded.